fix: connect wallet from settings and some refactors
