+++ web-marketplace/src/components/organisms/DashboardNavigation/DashboardNavigation.tsx
@@ -106,6 +106,7 @@ export const DashboardNavigation = ({
       hasOrders,
       hasCreditBatches,
       canEditOrg,
+      isOrganizationDashboard,
     ],
   );
