+++ web-marketplace/src/legacy-pages/Dashboard/Dashboard.Settings.tsx
@@ -1,17 +1,15 @@
 import { useLingui } from '@lingui/react';
+import { useSetAtom } from 'jotai';
 
 import { useAuth } from 'lib/auth/auth';
-import { useWallet } from 'lib/wallet/wallet';
-import { WalletType } from 'lib/wallet/walletsConfig/walletsConfig.types';
 
-import { AccountConnectWalletModal } from 'components/organisms/AccountConnectWalletModal/AccountConnectWalletModal';
-import { ConnectWalletFlow } from 'components/organisms/ConnectWalletFlow/ConnectWalletFlow';
-import { useLoginData } from 'components/organisms/LoginButton/hooks/useLoginData.legacy';
 import { useEmailConfirmationData } from 'components/organisms/LoginFlow/hooks/useEmailConfirmationData';
 import { UserAccountSettings } from 'components/organisms/UserAccountSettings/UserAccountSettings';
 import { EMAIL_ADDED } from 'components/organisms/UserAccountSettings/UserAccountSettings.constants';
 import { WalletProviderInfo } from 'components/organisms/UserAccountSettings/UserAccountSettings.types';
 
+import { useDashboardContext } from './Dashboard.context';
+import { shouldRedirectToCreateOrgAtom } from './Dashboard.store';
 import { useSocialProviders } from './hooks/useSocialProviders';
 import { useVerifyToken } from './hooks/useVerifyToken';
 
@@ -19,14 +17,10 @@ export const DashboardSettings = () => {
   const { _ } = useLingui();
   const { activeAccount, privActiveAccount } = useAuth();
   const hasKeplrAccount = !!activeAccount?.addr;
-  const { connect } = useWallet();
-  const {
-    isModalOpen,
-    modalState,
-    onButtonClick,
-    onModalClose,
-    walletsUiConfig,
-  } = useLoginData({});
+  const { onConnectWalletClick } = useDashboardContext();
+  const setShouldRedirectToCreateOrg = useSetAtom(
+    shouldRedirectToCreateOrgAtom,
+  );
 
   // Social providers
   const socialProviders = useSocialProviders();
@@ -48,7 +42,12 @@ export const DashboardSettings = () => {
   // Keplr account
   const walletProviderInfo: WalletProviderInfo = hasKeplrAccount
     ? { address: String(activeAccount?.addr) }
-    : { connect: onButtonClick };
+    : {
+        connect: () => {
+          setShouldRedirectToCreateOrg(false); // reset any org creation redirect
+          onConnectWalletClick();
+        },
+      };
 
   const emailConfirmationData = useEmailConfirmationData({
     emailConfirmationText: _(EMAIL_ADDED),
@@ -64,27 +63,6 @@ export const DashboardSettings = () => {
         walletProvider={walletProviderInfo}
         custodialAddress={activeAccount?.custodialAddress}
         emailConfirmationData={emailConfirmationData}
-      />
-      <AccountConnectWalletModal
-        open={isModalOpen}
-        onClose={onModalClose}
-        wallets={[
-          {
-            ...walletsUiConfig[0],
-            onClick: () => {
-              connect &&
-                connect({
-                  walletType: WalletType.Keplr,
-                  doLogin: false,
-                });
-            },
-          },
-        ]}
-        state={modalState}
-      />
-      <ConnectWalletFlow
-        isConnectModalOpened={isModalOpen}
-        onConnectModalClose={onModalClose}
       />
     </div>
   );
